changed ProjectPanel to only accept numeric values for Query Id field

+++ connectors/connector-jira/src/main/java/com/taskadapter/connector/jira/JqlBuilder.java
@@ -5,7 +5,7 @@ public final class JqlBuilder {
         return "project = " + projectKey;
     }
 
-    public static String findIssuesByProjectAndFilterId(String projectKey, int filterId) {
+    public static String findIssuesByProjectAndFilterId(String projectKey, long filterId) {
         return "project = " + projectKey + " AND filter = " + filterId;
     }
 }
